Added confirmation to article delete function

+++ public/js/article.js
@@ -3,17 +3,19 @@ const delButtonHandler = async (event) => {
   if (event.target.hasAttribute('data-id')) {
     const id = event.target.getAttribute('data-id');
 
-    console.log(`*****************************${id}`);
-
-    const response = await fetch(`/api/article/${id}`, {
-      method: 'DELETE',
-    });
-
-    if (response.ok) {
-      document.location.replace('/');
-    } else {
-      alert('Failed to delete article');
-    }
+    if (confirm('Are you sure you want to delete this article?')) {
+      console.log('Article deleted.');
+      const response = await fetch(`/api/article/${id}`, {
+        method: 'DELETE',
+      });
+      if (response.ok) {
+        document.location.replace('/');
+      } else {
+        alert('Failed to delete article');
+      }
+    } 
+  } else {
+    console.log('The article was not deleted.');
   }
 };
